backend/requirements.txt
asgiref==3.7.2
Django==4.1.7
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
djangorestframework==3.14.0
django-environ
gunicorn
psycopg2-binary==2.9.3